[231012] Linear/Logistic Regression, Naive Bayes
This is the assignment of Linear/Logistic Regression and Naive Bayes. 
In this session, I processed Titanic Dataset(logistic regression) to predict whether I would survive or not, and Iris dataset to predict the new input data in which class to be classified according to given features using machine learning algorithm.